Make Storage class for more OOP
