+++ src/main/java/YapPal.java
@@ -16,10 +16,6 @@ public class YapPal {
         TERMINATE,
     }
 
-    // operation constants
-    private static final int MAX_LIST_LEN = 100;
-    private static ArrayList<Task> tasks = new ArrayList<Task>(YapPal.MAX_LIST_LEN);
-
     public static void main(String[] args) {
         // initialisation
         Ui.printIntro();
@@ -141,44 +137,10 @@ public class YapPal {
         }
         Task targetedTask = YapPal.tasks.get(ptr - 1);
         YapPal.tasks.remove(ptr - 1);
+        YapPal.save();
         Ui.printMsg(
             "OK, I've removed this task: \n" +
             targetedTask
         );
-    }
-
-    private static void save() {
-        try {
-            FileWriter saveFileWriter = new FileWriter(YapPal.SAVE_DIRECTORY);
-            for (int i = 0; i < tasks.size(); ++i) {
-                saveFileWriter.write(tasks.get(i).saveString() + "\n");
-            }
-            saveFileWriter.close();
-        } catch (IOException error) {
-            Ui.printMsg(error.toString());
-        }
-    }
-
-    private static ArrayList<Task> load() throws YapPalException {
-        File saveFile = new File(YapPal.SAVE_DIRECTORY);
-        ArrayList<Task> tasks = new ArrayList<>(YapPal.MAX_LIST_LEN);
-        try {
-            Scanner saveReader = new Scanner(saveFile);
-            while (saveReader.hasNextLine()) {
-                String command = saveReader.nextLine();
-                Task task = determineTask(command);
-                tasks.add(task);
-            }
-            saveReader.close();
-        } catch (FileNotFoundException exception) {
-            System.out.println("Save file not found - creating new save");
-            try {
-                saveFile.getParentFile().mkdir();
-                saveFile.createNewFile();
-            } catch (IOException fileCreateException) {
-                System.out.println("An error occurred");
-            }
-        }
-        return tasks;
     }
 }
